2019/src/test/java/aoc/Day22Test.java
package aoc;

import org.junit.jupiter.api.*;

import java.io.IOException;
import java.math.BigInteger;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Paths;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day22Test {
    private static String input;
    private Day22 day;

    @BeforeAll
    public static void init() throws IOException {
        input = Files.readString(Paths.get("src/main/resources/day22.txt"), StandardCharsets.UTF_8);
    }

    @BeforeEach
    public void setup() {
        day = new Day22();
    }

    @Nested
    @DisplayName("Part 1")
    class Part1Tests {
        @Test
        void testSolution() {
            BigInteger result = day.part1(input);
            assertEquals(BigInteger.valueOf(4649L), result);
        }
    }

    @Nested
    @DisplayName("Part 2")
    class Part2Tests {
        @Test
        void testSolution() {
            BigInteger result = day.part2(input);
            assertEquals(BigInteger.valueOf(68849657493596L), result);
        }
    }
}